Fix another sprintf warning in example.

+++ asio/src/examples/cpp11/chat/chat_message.hpp
@@ -79,7 +79,7 @@ public:
   void encode_header()
   {
     char header[header_length + 1] = "";
-    std::sprintf(header, "%4d", body_length_);
+    std::sprintf(header, "%4d", static_cast<int>(body_length_));
     std::memcpy(data_, header, header_length);
   }
